bbox_compute: activated version

+++ gaussian_model/gaussian_model.py
@@ -24,11 +24,11 @@ for n in range(N_max):
 
 def compute_relocation_cuda(opacity_old, scale_old, N):
     N.clamp_(min=1, max=N_max-1)
-    return compute_relocation(opacity_old, scale_old, N, binoms, N_max)
+    return compute_relocation(opacity_old, scale_old, N.int(), binoms, N_max)
 
 def simple_compute_relocation_cuda(opacity_old, scale_old, N):
     N.clamp_(min=1, max=N_max-1)
-    return simple_compute_relocation(opacity_old, scale_old, N, binoms, N_max)
+    return simple_compute_relocation(opacity_old, scale_old, N.int(), binoms, N_max)
 
 class GaussianModel:
     def setup_functions(self):
